Split SvgPathSegment
